requirements.txt
# Bouncer Dependencies

# Core
claude-agent-sdk>=0.1.12
watchdog>=3.0.0
pyyaml>=6.0.0
aiohttp>=3.9.0

# Optional linters and formatters (install as needed)
# pylint>=3.0.0
# black>=23.0.0
# eslint (requires npm)
# prettier (requires npm)

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
